Move periodic progress logging to DockerInterface for all image pulls
Move the periodic progress logging from HomeAssistantCore.install() down
to DockerInterface.install() so that all container image pulls (Core,
add-ons, plugins) benefit from visibility during long pulls. Every 15
seconds, a log message reports the current pull progress percentage or
indicates the pull is still in progress if no progress data is available.
